requirements.txt
autopep8==1.7.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Toastr==0.5.7
Flask-WTF==1.0.1
gunicorn==20.1.0
harperdb==0.4.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.9.1
pycparser==2.21
python-dotenv==0.20.0
requests==2.28.1
toml==0.10.2
urllib3==1.26.11
Werkzeug==2.2.2
whitenoise==6.2.0
WTForms==3.0.1
